fixed: fixed images size
