+++ src/shared/components/custom-slider/CustomSlider.module.scss
