+++ src/shared/components/custom-slider/CustomSlider.tsx
