+++ src/shared/components/expandable-text/index.ts
@@ -1,0 +1,1 @@
+export { ExpandableText } from './ui/ExpandableText'
