+++ src/shared/components/expandable-text/ui/ExpandableText.module.scss
@@ -1,0 +1,22 @@
+.text {
+  width: 100%;
+
+  font-size: 14px;
+  font-weight: 400;
+  font-style: normal;
+  line-height: 24px;
+  color: var(--light-100, #fff);
+}
+
+.toggleButton {
+  cursor: pointer;
+
+  margin-left: 7px;
+
+  font-size: 14px;
+  font-weight: 400;
+  font-style: normal;
+  line-height: 24px;
+  color: var(--primary-500, #397df6);
+  text-decoration-line: underline;
+}
