+++ src/shared/components/public-profile-card/index.ts
@@ -1,0 +1,1 @@
+export { PublicProfileCard } from './ui/PublicProfileCard'
