+++ src/shared/components/public-profile-card/ui/PublicProfileCard.module.scss
@@ -1,0 +1,98 @@
+.container {
+  flex-shrink: 0;
+  width: 234px;
+  margin-top: 36px;
+
+  .wrapper {
+    display: flex;
+    width: 234px;
+    height: 200px;
+  }
+}
+
+.sticky {
+  z-index: 11;
+  align-self: flex-end;
+  min-height: 180px;
+  background-color: var(--color-dark-500);
+}
+
+.mainImage {
+  flex-shrink: 0;
+
+  width: 234px;
+  height: 240px;
+
+  background: var(--color-dark-500, #171717);
+  border-radius: 1px;
+}
+
+.profileImg {
+  flex-shrink: 0;
+
+  width: 36px;
+  height: 36px;
+  margin-right: 18px;
+
+  border-radius: 50%;
+}
+
+.mainInfo {
+  display: flex;
+  align-items: center;
+  margin: 12px 0;
+}
+
+.timeInfo {
+  color: var(--light-900, #8d9094);
+}
+
+.description {
+  .text {
+    width: 100%;
+
+    font-size: 14px;
+    font-weight: 400;
+    font-style: normal;
+    line-height: 24px;
+    color: var(--light-100, #fff);
+  }
+
+  .toggleButton {
+    cursor: pointer;
+
+    margin-left: 7px;
+
+    font-size: 14px;
+    font-weight: 400;
+    font-style: normal;
+    line-height: 24px;
+    color: var(--primary-500, #397df6);
+    text-decoration-line: underline;
+  }
+}
+
+@media (width > 1280px) {
+  .container {
+    flex-shrink: 0;
+    width: 360px;
+    margin-top: 36px;
+
+    .wrapper {
+      display: flex;
+      width: 360px;
+      height: 160px;
+    }
+
+    .sticky {
+      min-width: 360px;
+      min-height: 150px;
+    }
+  }
+
+  .mainImage {
+    flex-shrink: 0;
+    width: 360px;
+    height: 360px;
+  }
+}
