+++ src/widgets/publicProfileCardsList/ui/PublicProfileCardsList.module.scss
@@ -9,272 +9,3 @@
     height: 200px;
   }
 }
-
-.sticky {
-  z-index: 11;
-  align-self: flex-end;
-  min-height: 180px;
-  background-color: var(--color-dark-500);
-}
-
-.mainImage {
-  flex-shrink: 0;
-
-  width: 234px;
-  height: 240px;
-
-  background: var(--color-dark-500, #171717);
-  border-radius: 1px;
-}
-
-.profileImg {
-  flex-shrink: 0;
-
-  width: 36px;
-  height: 36px;
-  margin-right: 18px;
-
-  border-radius: 50%;
-}
-
-.mainInfo {
-  display: flex;
-  align-items: center;
-  margin: 12px 0;
-}
-
-.timeInfo {
-  color: var(--light-900, #8d9094);
-}
-
-.description {
-  .text {
-    width: 100%;
-
-    font-size: 14px;
-    font-weight: 400;
-    font-style: normal;
-    line-height: 24px;
-    color: var(--light-100, #fff);
-  }
-
-  .toggleButton {
-    cursor: pointer;
-
-    margin-left: 7px;
-
-    font-size: 14px;
-    font-weight: 400;
-    font-style: normal;
-    line-height: 24px;
-    color: var(--primary-500, #397df6);
-    text-decoration-line: underline;
-  }
-}
-
-@media (640px < width <=768px) {
-  .container {
-    flex-shrink: 0;
-    width: 120px;
-    margin-top: 36px;
-
-    .wrapper {
-      display: flex;
-      width: 120px;
-      height: 200px;
-    }
-  }
-
-  .sticky {
-    align-self: flex-end;
-    min-height: 180px;
-    background-color: var(--color-dark-500);
-  }
-
-  .mainImage {
-    flex-shrink: 0;
-
-    width: 120px;
-    height: 120px;
-
-    background: var(--dark-500, #171717);
-    border-radius: 1px;
-  }
-
-  .profileImg {
-    flex-shrink: 0;
-
-    width: 36px;
-    height: 36px;
-    margin-right: 8px;
-
-    border-radius: 50%;
-  }
-
-  .mainInfo {
-    display: flex;
-    align-items: center;
-    margin: 12px 0;
-
-    .profileName {
-      margin-right: 4px;
-
-      font-size: 10px;
-      font-weight: 400;
-      font-style: normal;
-      line-height: 16px;
-      text-align: left;
-    }
-  }
-
-  .timeInfo {
-    color: var(--light-900, #8d9094);
-  }
-
-  .description {
-    .text {
-      width: 100%;
-
-      font-size: 10px;
-      font-weight: 700;
-      font-style: normal;
-      line-height: 16px;
-      color: var(--light-100, #fff);
-    }
-
-    .toggleButton {
-      cursor: pointer;
-
-      margin-left: 7px;
-
-      font-size: 10px;
-      font-weight: 700;
-      font-style: normal;
-      line-height: 16px;
-      color: var(--primary-500, #397df6);
-      text-decoration-line: underline;
-    }
-  }
-}
-
-@media (640px > width) {
-  .container {
-    flex-shrink: 0;
-    max-width: 234px;
-    margin-top: 36px;
-
-    .wrapper {
-      display: flex;
-      height: 200px;
-    }
-  }
-
-  .sticky {
-    align-self: flex-end;
-    background-color: var(--color-dark-500);
-  }
-
-  .mainImage {
-    flex-shrink: 0;
-
-    width: 234px;
-    height: 240px;
-
-    background: var(--dark-500, #171717);
-    border-radius: 1px;
-  }
-
-  .profileImg {
-    flex-shrink: 0;
-
-    width: 36px;
-    height: 36px;
-    margin-right: 18px;
-
-    border-radius: 50%;
-  }
-
-  .mainInfo {
-    display: flex;
-    align-items: center;
-    margin: 12px 0;
-  }
-
-  .timeInfo {
-    color: var(--light-900, #8d9094);
-  }
-
-  .description {
-    .text {
-      width: 100%;
-
-      font-size: 14px;
-      font-weight: 400;
-      font-style: normal;
-      line-height: 24px;
-      color: var(--light-100, #fff);
-    }
-
-    .toggleButton {
-      cursor: pointer;
-
-      margin-left: 7px;
-
-      font-size: 14px;
-      font-weight: 400;
-      font-style: normal;
-      line-height: 24px;
-      color: var(--primary-500, #397df6);
-      text-decoration-line: underline;
-    }
-  }
-}
-
-@media (1024px > width > 768px) {
-  .container {
-    flex-shrink: 0;
-    width: 160px;
-    margin-top: 36px;
-
-    .wrapper {
-      display: flex;
-      width: 160px;
-      height: 240px;
-    }
-
-    .sticky {
-      min-height: 220px;
-    }
-  }
-
-  .mainImage {
-    flex-shrink: 0;
-    width: 160px;
-    height: 160px;
-  }
-}
-
-@media (width > 1536px) {
-  .container {
-    flex-shrink: 0;
-    width: 360px;
-    margin-top: 36px;
-
-    .wrapper {
-      display: flex;
-      width: 360px;
-      height: 160px;
-    }
-
-    .sticky {
-      min-width: 360px;
-      min-height: 150px;
-    }
-  }
-
-  .mainImage {
-    flex-shrink: 0;
-    width: 360px;
-    height: 360px;
-  }
-}
