+++ src/widgets/registeredUsersCounter/ui/RegisteredUsers.module.scss
@@ -26,8 +26,8 @@
   }
 
   .counter li {
-    padding: 3px;
     width: 24px;
+    padding: 3px;
 
     font-family: var(--font-family-primary);
     font-size: 18px;
